moved OT from detector to metric
