+++ aif360/detectors/__init__.py
@@ -1,3 +1,2 @@
 from aif360.detectors.mdss.MDSS import MDSS
 from aif360.detectors.mdss_detector import bias_scan
-from aif360.detectors.ot_detector import ot_bias_scan
