+++ aif360/sklearn/detectors/__init__.py
@@ -2,9 +2,7 @@
 Methods for detecting subsets for which a model or dataset is biased.
 """
 from aif360.sklearn.detectors.detectors import bias_scan
-from aif360.sklearn.detectors.detectors import ot_bias_scan
 
 __all__ = [
     'bias_scan',
-    'ot_bias_scan',
 ]
