+++ aif360/sklearn/detectors/detectors.py
@@ -1,71 +1,9 @@
-from aif360.detectors.ot_detector import ot_bias_scan
 from aif360.detectors import bias_scan
 from aif360.detectors.mdss.ScoringFunctions import ScoringFunction
 
 from typing import Union
 import pandas as pd
 import numpy as np
-
-def ot_bias_scan(
-    y_true: pd.Series,
-    y_pred: Union[pd.Series, pd.DataFrame],
-    prot_attr: pd.Series = None,
-    pos_label: Union[str, float] = None,
-    overpredicted: bool = True,
-    scoring: str = "Wasserstein1",
-    num_iters: int = 100,
-    penalty: float = 1e-17,
-    mode: str = "ordinal",
-    cost_matrix: np.ndarray=None,
-    **kwargs,
-):
-    """Normalize and calculate Wasserstein distance between groups defined by `prot_attr` in `y_true` and `y_pred`.
-
-    Args:
-        y_true (pd.Series): ground truth (correct) target values.
-        y_pred (pd.Series, pd.DataFrame): estimated target values.
-            If `mode` is nominal, must be a `pd.DataFrame` with columns containing predictions for each nominal class,
-                or list of corresponding column names in `data`.
-            If `None`, model is assumed to be a dummy model that predicts the mean of the targets
-                or 1/(number of categories) for nominal mode.
-        sensitive_attribute (pd.Series): sensitive attribute values.
-            If `None`, assume all samples belong to the same protected group.
-        pos_label(str, float, optional): Either "high", "low" or a float value if the mode in [binary, ordinal, or continuous].
-                If float, value has to be the minimum or the maximum in the ground_truth column.
-                Defaults to high if None for these modes.
-                Support for float left in to keep the intuition clear in binary classification tasks.
-                If `mode` is nominal, favorable values should be one of the unique categories in the ground_truth.
-                Defaults to a one-vs-all scan if None for nominal mode.
-        overpredicted (bool, optional): flag for group to scan for.
-            `True` scans for overprediction, `False` scans for underprediction.
-        scoring (str or class): only 'Wasserstein1'
-        num_iters (int, optional): number of iterations (random restarts) for EMD. Should be positive.
-        penalty (float, optional): penalty term. Should be positive. The penalty term as with any regularization parameter
-            may need to be tuned for a particular use case. The higher the penalty, the higher the influence of entropy regualizer.
-        mode: one of ['binary', 'continuous', 'nominal', 'ordinal']. Defaults to binary.
-                In nominal mode, up to 10 categories are supported by default.
-                To increase this, pass in keyword argument max_nominal = integer value.
-        cost_matrix (np.ndarray): cost matrix for the Wasserstein distance. Defaults to absolute difference between samples.
-
-    Returns:
-        ot.emd2 (float, dict): Earth mover's distance or dictionary of optimal transports for each of option of classifier
-
-    Raises:
-        ValueError: if `mode` is 'binary' but `ground_truth` contains less than 1 or more than 2 unique values.
-    """
-    return ot_bias_scan(
-        ground_truth=y_true,
-        classifier=y_pred,
-        prot_attr=prot_attr,
-        favorable_value=pos_label,
-        overpredicted=overpredicted,
-        scoring=scoring,
-        num_iters=num_iters,
-        penalty=penalty,
-        mode=mode,
-        cost_matrix=cost_matrix,
-        kwargs=kwargs
-    )
 
 def bias_scan(
     X: pd.DataFrame,
