+++ tests/test_emd_metric.py
@@ -1,6 +1,6 @@
 import numpy as np
 import pandas as pd
-from aif360.detectors.ot_detector import ot_bias_scan, _normalize, _transform, _evaluate
+from aif360.metrics.emd_metric import earth_movers_distance, _normalize, _transform, _evaluate
 from ot import emd2
 import unittest
 from unittest import TestCase
@@ -94,28 +94,28 @@ class TestEvaluate():
 class TestOtBiasScan(TestCase):
     def test_scoring_checked(self):
         with self.assertRaises(Exception):
-            ot_bias_scan(pd.Series(), pd.Series(), scoring="Bernoulli")
+            earth_movers_distance(pd.Series(), pd.Series(), scoring="Bernoulli")
     def test_scan_mode_checked(self):
         with self.assertRaises(Exception):
-            ot_bias_scan(pd.Series(), pd.Series(), mode="Wrong")
+            earth_movers_distance(pd.Series(), pd.Series(), mode="Wrong")
 
     def test_classifier_type_checked(self):
         with self.assertRaises(Exception):
-            ot_bias_scan(pd.Series(), pd.Series(), mode="nominal")
+            earth_movers_distance(pd.Series(), pd.Series(), mode="nominal")
         with self.assertRaises(Exception):
-            ot_bias_scan(pd.Series(), pd.Series(), mode="ordinal")
+            earth_movers_distance(pd.Series(), pd.Series(), mode="ordinal")
         with self.assertRaises(Exception):
-            ot_bias_scan(pd.Series(), pd.DataFrame(), mode="binary")
+            earth_movers_distance(pd.Series(), pd.DataFrame(), mode="binary")
         with self.assertRaises(Exception):
-            ot_bias_scan(pd.Series(), pd.DataFrame(), mode="continuous")
+            earth_movers_distance(pd.Series(), pd.DataFrame(), mode="continuous")
 
     def test_binary_nuniques_checked(self):
         with self.assertRaises(Exception):
-            ot_bias_scan(pd.Series([1,2,3], pd.Series(), mode='binary'))
+            earth_movers_distance(pd.Series([1,2,3], pd.Series(), mode='binary'))
     
     def test_cost_matrix_type_checked(self):
         with self.assertRaises(TypeError):
-            ot_bias_scan(pd.Series(), pd.Series(), cost_matrix=pd.DataFrame())
+            earth_movers_distance(pd.Series(), pd.Series(), cost_matrix=pd.DataFrame())
 
     def test_cost_matrix_passed_correctly(self):
         p = pd.Series([50, 100, 150])
@@ -124,7 +124,7 @@ class TestOtBiasScan(TestCase):
                       [20, 40, 15, 30, 30], 
                       [30, 35, 40, 55, 25]])
         expected = _evaluate(p, q, cost_matrix=C)
-        actual = ot_bias_scan(p, q, cost_matrix=C, mode="continuous")
+        actual = earth_movers_distance(p, q, cost_matrix=C, mode="continuous")
         assert expected == actual
 
     def test_favorable_value_checked(self):
@@ -132,25 +132,25 @@ class TestOtBiasScan(TestCase):
         q = pd.Series([25, 115, 60, 30, 70])
         fav = 4
         with self.assertRaises(ValueError):
-            ot_bias_scan(p, q, favorable_value=fav)
+            earth_movers_distance(p, q, favorable_value=fav)
 
     def test_nominal_classifier_shape_checked(self):
         p = pd.Series([0,0,1,1])
         q = pd.DataFrame([0.5,0.5,0.5,0.5])
         with self.assertRaises(ValueError):
-            ot_bias_scan(p, q, mode='nominal')
+            earth_movers_distance(p, q, mode='nominal')
     
     def test_ordinal_classifier_shape_checked(self):
         p = pd.Series([0,0,1,1])
         q = pd.DataFrame([0.5,0.5,0.5,0.5])
         with self.assertRaises(ValueError):
-            ot_bias_scan(p, q, mode='ordinal')
+            earth_movers_distance(p, q, mode='ordinal')
 
     def test_nominal(self):
         p = pd.Series([0,0,1,1])
         q = pd.DataFrame([[0.5,0.5],[0.5,0.5],[0.5,0.5],[0.5,0.5]])
         expected = {cl: _evaluate(p, q[cl]) for cl in [0,1]}
-        actual = ot_bias_scan(p, q, mode="nominal")
+        actual = earth_movers_distance(p, q, mode="nominal")
         assert expected == actual
